requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-admin-persian-fonts==0.2
django-chartjs==2.3.0
django-jalali-date==1.0.2
django-tinymce==3.6.1
idna==3.4
jdatetime==4.1.1
Pillow==10.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
